% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.mode.R
\name{calculate.mode}
\alias{calculate.mode}
\title{Mode calculation}
\usage{
calculate.mode(v)
}
\arguments{
\item{v}{A vector of numeric numbers}
}
\value{
A single number corresponding to the mode of the list of numbers give as input
}
\description{
Calculate the mode value of a vector of numeric values.
}
\examples{
mode = calculate.mode(v = c(6, 8, 4, 845, 8, 5, 55, 84, 8, 84, 45, 5))

}
