% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest.regions.R
\name{closest.regions}
\alias{closest.regions}
\title{Find closets regions to reference regions.}
\usage{
closest.regions(
  reference.regions,
  reference.regions.table.name = "referenceRegions",
  target.regions,
  export.table.file = NULL,
  return.table = TRUE,
  collapse.regions = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{reference.regions}{A full path to a bed file or a data.frame in at least BED3 format with the regions to use as reference.}

\item{reference.regions.table.name}{A string with the name to use for the group reference regions. By default \code{"referenceRegions"}.}

\item{target.regions}{A full path to a bed file or a data.frame in at least BED3 format with the regions to uses as targets.}

\item{export.table.file}{A string with the full path for the file in which the table should be exported. By default \code{NULL}: not export.}

\item{return.table}{Logical value to define whether the output table should be returned. By default \code{TRUE}.}

\item{collapse.regions}{Logical value to define whether the partially overlapping regions should be collapsed or not. By default \code{FALSE}.}

\item{verbose}{Logical value to define whether messages should be printed. By default \code{TRUE}.}
}
\value{
The function returns a data.frame composed of a triplicated chr-start-end-name table for reference.region, upstream.region and downstream.region, respectively.
}
\description{
This tools return the closest upstream and downstream regions from a reference region.
}
