% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmyk.R
\name{cmyk}
\alias{cmyk}
\title{CMYK color converter}
\usage{
cmyk(C, M, Y, K)
}
\arguments{
\item{C}{Value in the 0-100 range for the Cyan component.}

\item{M}{Value in the 0-100 range for the Magenta component.}

\item{Y}{Value in the 0-100 range for the Yellow component.}

\item{K}{Value in the 0-100 range for the Key component.}
}
\value{
The result is a string for the color in hexadecimal scale, eg. "#FFFFFF".
}
\description{
Converts CMYK color values to hexadecimal color values
}
\examples{
color = cmyk(0, 0, 0, 0)

}
