% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CNV.data}
\alias{CNV.data}
\title{CNV data results example}
\format{
A data frame with 25 rows and 9 variables:
\describe{
  \item{\code{geneName}}{hypothetical gene symbols}
  \item{\code{patient_1 ... patient_N}}{hypothetical patients ID}
}
}
\source{
Simulated data
}
\usage{
CNV.data
}
\description{
Simulation of Copy Number Variation (CNV) analysis on a cohort of patients.
}
\keyword{datasets}
