% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.gradient.R
\name{color.gradient}
\alias{color.gradient}
\title{Gradient colors generation and assignment}
\usage{
color.gradient(values, colors = c("blue", "white", "red"), bins = 100)
}
\arguments{
\item{values}{A numeric vector containing the values to which a color must be assigned (NAs and NaN will be converted to 0).}

\item{colors}{A string vector with the colors, in the wished order, that have to be used to generated the shadows. By default \code{c("blue", "white", "red")}.}

\item{bins}{An atomic integer value to define the total number of bins/steps in which the gradient should be dived.}
}
\value{
A vector containing the assigned colors corresponding to each element of \code{values}.
}
\description{
Give a vector of colors generates a finite number of shadows that will be assigned to a numeric vector depending on the value of each element.
}
