% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sequence.R
\name{convert_sequence}
\alias{convert_sequence}
\title{Nucleic acid sequences converter.}
\usage{
convert_sequence(sequence = NULL, mode = "not specified", nucleic.acid = "DNA")
}
\arguments{
\item{sequence}{A string containing the sequence to be converted. By default \code{NULL}, it returns an help for the mode.}

\item{mode}{A string value to define the modality of convertion. Possible options: \cr - Reverse complement = revComp | RC | rc | reverseComplement \cr - Reverse            = rev     | R  | r  | reverse \cr - Complement         = comp    | C  | c  | complement. \cr By default \code{"not specified"}, it returns an help for the mode.}

\item{nucleic.acid}{A string to define the type of nucleic acid to which the input sequence belongs. Available options "DNA", default value, or "RNA".}
}
\value{
It returns a string with the converted sequence.
}
\description{
Obtains de complementary, reverse complementary or the reverse of a DNA/RNA sequence.
}
\examples{
convert_sequence(sequence = "AATTTCCCGTCGAT",
                 mode = "reverse",
                 nucleic.acid = "DNA")

}
