% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.to.list.R
\name{data.frame.to.list}
\alias{data.frame.to.list}
\title{Data frame conversion to a list of columns.}
\usage{
data.frame.to.list(x)
}
\arguments{
\item{x}{A data.frame to be converted}
}
\value{
A list of vectors in which each element is a column of input the data.frame.
}
\description{
Converts each column of a data.frame in a element of a list with the corresponding name of the original column. Useful for further use in functions such as purrr::pmap().
}
\examples{
data.frame.to.list(mtcars)

}
