% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.summary.R
\name{data.summary}
\alias{data.summary}
\title{Statistical data summary generator}
\usage{
data.summary(data, variable, group.names)
}
\arguments{
\item{data}{Input data.frame to be analyzed.}

\item{variable}{A string with the name of the column to be analyzed.}

\item{group.names}{A string with the name of the column indicating the groups.}
}
\value{
It returns a list that is a combination of the lists in the input list. \cr If the list is not a nested list of list the original input is returned.
}
\description{
Produces a table with a summary of the statistics for a specific column of an input data.frame by a group of values defined by a group defined by another column.
}
\examples{
data.summary(data = mtcars, variable = "mpg", group.names = "disp")

}
