% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.status.R
\name{DE.status}
\alias{DE.status}
\title{Differential Expression status calculator for RNA-seq data}
\usage{
DE.status(
  log2FC,
  p.value.adjusted,
  FC_threshold = 1.5,
  FC_NoResp_left = 0.9,
  FC_NoResp_rigth = NULL,
  p.value_threshold = 0.05,
  low.FC.status.label = "DOWN",
  high.FC.status.label = "UP",
  unresponsive.label = "NoResp",
  null.label = "NULL"
)
}
\arguments{
\item{log2FC}{Numeric vector of log2(fold change expression) values.}

\item{p.value.adjusted}{Numeric vector of p-values. Use of adjusted p-values is recommended.}

\item{FC_threshold}{Value of the threshold to use for the fold change expression to define differentially expressed genes, expressed as linear value. By default 1.5 and by consequence 1/1.5.}

\item{FC_NoResp_left}{Value of the threshold to use for the fold change expression to define unresponsive genes when \code{FC < 1}, expressed as linear value. By default 0.9. If \code{NULL} it will be calculated symmetrically from \code{FC_NoResp_rigth} as 1/\code{FC_NoResp_rigth}.}

\item{FC_NoResp_rigth}{Value of the threshold to use for the fold change expression to define unresponsive genes when \code{FC > 1}, expressed as linear value. By default 1.1. If \code{NULL} it will be calculated symmetrically from \code{FC_NoResp_left} as 1/\code{FC_NoResp_left}.}

\item{p.value_threshold}{Value of the threshold to use for the p-values to define differentially expressed genes, expressed as linear value. By default 0.05.}

\item{low.FC.status.label}{String to define the label indicating the differentially expressed genes with a \code{FoldChange < FC_threshold}.}

\item{high.FC.status.label}{String to define the label indicating the differentially expressed genes with a \code{FoldChange > FC_threshold}.}

\item{unresponsive.label}{String to define the label indicating the unresponsive genes identified as \code{FC_NoResp_left < FoldChange < FC_NoResp_rigth} and \code{p.value > p.value.threshold}.}

\item{null.label}{String to define the label indicating the null genes.}
}
\value{
It returns a vector containing the differential expression status for each original value in the same order used in the input.
}
\description{
Defines the differential expression status of genes from RNA-seq data depending on fold change expression and adjusted p-value.
}
