% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.matrix.R
\name{density.matrix}
\alias{density.matrix}
\title{Density matrix builder}
\usage{
\method{density}{matrix}(
  mode,
  regions.list,
  samples.list,
  region.names = NULL,
  sample.names = NULL,
  sort.regions.coordinates = FALSE,
  reference.point = "center",
  reference.point.label = NULL,
  upstream = 500,
  downstream = 500,
  body.length = 1000,
  missing.data.as.zero = FALSE,
  bin.size = 10,
  binning.operation = "mean"
)
}
\arguments{
\item{mode}{A string indicating the method for the matrix computation:
\itemize{
  \item \code{scale-regions} all regions in the BED file are stretched or shrunken to the length (in bases) indicated by the user (\code{body.length});
  \item \code{reference-point} the matrix will be performed on the range -upstream+downstream from the indicated reference point (center, TSS, TES).
 }}

\item{regions.list}{A string vector with a list of full paths to bed files or list of data.frames in at least BED3 format (eg. generated by \link{build.bed}).}

\item{samples.list}{A string vector with a list of full paths to bigWig files.}

\item{region.names}{A string vector with the names of the regions. If \code{NULL} or of length lower than the number of regions the names will be assigned using the basename of the file if a path is provided otherwise "region_<order number>". By default \code{NULL}.}

\item{sample.names}{A string vector with the names of the samples. If \code{NULL} or of length lower than the number of samples the names will be assigned using the basename of the file. By default \code{NULL}.}

\item{sort.regions.coordinates}{Logical value to define whether the output matrix should contain the regions sorted by genomic location for each region group (sorted by \link{sort.bed}). By default \code{FALSE}.}

\item{reference.point}{The reference point for the matrix generation could be either the region start (\code{"TSS"}), the region end (\code{"TES"}) or the \code{"center"} of the region. By default \code{"center"}.}

\item{reference.point.label}{A single string with the label for the reference point that could be used for the plots.}

\item{upstream}{Distance, in bases (bp), upstream of the reference-point, in "reference-point" mode, or the region start, in "scale-regions" mode. By default \code{500}.}

\item{downstream}{Distance, in bases (bp), downstream of the reference-point, in "reference-point" mode, or the region start, in "scale-regions" mode. By default \code{500}.}

\item{body.length}{Distance, in bases (bp), to which all regions will be fit. By default: \code{1000}.}

\item{missing.data.as.zero}{A logical value to define whether missing data (NAs) should be treated as zeros. By default \code{FALSE}.}

\item{bin.size}{Length, in bases (bp), of the non-overlapping bins for averaging the score over the regions length. By default \code{10}.}

\item{binning.operation}{A single string to define the type of statistic that should be used over the bin size range. The options are: "mean", "median", "sum". By default \code{"mean"}.}
}
\value{
The function returns a named list containing:
\itemize{
  \item \code{metadata} data.frame with the parameters used to build the matrix;
  \item \code{matrix.data} data.frame with the computed scores;
  \item \code{original.file.path} with the string: "Matrix generated by Rseb::density.matrix()".
 }
 This list can be passed as it is to the functions \link{plot.density.profile}, \link{plot.density.summary} and, \link{plot.density.differences}.
}
\description{
A function (completely in R) that generates a matrix given a list of regions (.bed files) and signals (.bigWig files) alternative (even though more time consuming) to \link{computeMatrix.deeptools}. The output can be passed as it is to the functions \link{plot.density.profile}, \link{plot.density.summary} and, \link{plot.density.differences}.
}
