% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{density_plot}
\alias{density_plot}
\title{Plot density signal of NGS data.}
\usage{
density_plot(
  samples,
  scores,
  positions,
  variance_scores,
  xlab = "Distance from regions center [bp]",
  ylab = "Average density signal",
  line_type = "solid",
  y_lim = NULL,
  x_lim = NULL,
  x_intercept = 0,
  colors = c("blue", "red", "purple", "orange", "green"),
  title = "Density profile",
  text_size = 12,
  variance = T,
  print_plot = F,
  line_width = 1,
  variance_opacity = 0.25
)
}
\arguments{
\item{samples}{A character vector containing the samples list.}

\item{scores}{A numeric vector containing the scores for the Y-axis.}

\item{positions}{A numeric vector containing the position for the X-axis.}

\item{variance_scores}{A numeric vector containing the variance/error value at each position.}

\item{xlab}{A string containing the label for the X-axis. By default "Distance from regions center [bp]".}

\item{ylab}{A string containing the label for the Y-axis. By default "Average density signal".}

\item{line_type}{Vector to define each line type. Both numeric and string codes are accepted. if only one element is given this will be applied to all the lines. By default "solid". \cr Example 1: \code{c("solid", "dashed")}. \cr Example 2: \code{c(1, 2)}}

\item{y_lim}{List of numeric vectors with two elements each to define the range of the Y-axis. To set only one side use NA for the side to leave automatic. If only one range is given this one will be applied to all the plots. By default \code{NULL}, the range will be defined automatically. \cr Example \code{list(c(0, 20), c(NA, 30), c(0, NA), c(NA, NA))}.,}

\item{x_lim}{List of numeric vectors with two elements each to define the range of the X-axis. To set only one side use NA for the side to leave automatic. If only one range is given this one will be applied to all the plots. By default \code{NULL}, the range will be defined automatically. \cr Example \code{list(c(0, 20), c(NA, 30), c(0, NA), c(NA, NA))}.,}

\item{x_intercept}{A vector indicating the X intercepts for the vertical lines. By default 0.}

\item{colors}{Vector to define the colors of the lines. If only one value is provided or the number of values are less then the required ones only the first value will be used. All standard R.colors values are accepted. By default \code{c("blue", "red", "purple", "orange", "green")}.}

\item{title}{A string containing the label for the X-axis. By default "Density profile".}

\item{text_size}{Numeric value to define the size of the text for the labels of all the plots. By default 12.}

\item{variance}{Logic value to define whether to plot the error/variance around the signal. By default \code{TRUE}.}

\item{print_plot}{Logic value to define whether to print the plot once generated or not. By default \code{FALSE}.}

\item{line_width}{Numeric value to define the line width for all the plots. By default 1.,}

\item{variance_opacity}{Numeric value to define the alpha/transparency of the error/variance. By default 0.25. Parameter considered only when \code{variance = TRUE)}.}
}
\value{
Returns a plot in ggplot2 format.
}
\description{
Plots the density profile of NGS data (e.g. ChIP-seq, ATAC-seq, MeDIP-seq, etc.). Used by the function \code{\link{plot.density.profile}}.
}
