% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doughnut.R
\name{doughnut}
\alias{doughnut}
\title{Donut/Doughnut plot}
\usage{
doughnut(
  x,
  labels = as.character(x),
  edges = 200,
  outer.radius = 0.8,
  inner.radius = 0.4,
  clockwise = FALSE,
  init.angle = if (clockwise) 90 else 0,
  density = NULL,
  angle = 45,
  col = NULL,
  border = FALSE,
  lty = NULL,
  main = NULL,
  ...
)
}
\arguments{
\item{x}{A vector containing the values to be plotted.}

\item{labels}{A string vector for the labels of the different sectors. By default as.character(x).}

\item{edges}{Number of edges of the shape. By default 200.}

\item{outer.radius}{Fraction of the area to dedicate to the outer circle. By default 0.8.}

\item{inner.radius}{Fraction of the area to dedicate to the inner circle. By default 0.4.}

\item{clockwise}{Logic value to define whether the values should be plotted in clockwise sense. By default \code{FALSE}.}

\item{init.angle}{Numeric value to define the starting angle for the data. By default if \code{clockwise = TRUE} 90, otherwise 0.}

\item{density}{A vector or single number to define de density of the lines in the filling color of each value plotted. By default \code{NULL}.}

\item{angle}{A vector or single number to define de angle of the lines in the filling color of each value plotted. By default 45.}

\item{col}{A vector of R standard colors for each value to be plotted. By default \code{NULL}.}

\item{border}{Logic value to define whether plot the border of the sectors. By default \code{FALSE}.}

\item{lty}{Numeric value to define the type of line for the borders. By default \code{NULL}.}

\item{main}{String to set the title of the plot. By default \code{NULL}.}
}
\description{
Generation of a donut/doughnut plot (equivalent of a pie chart)
}
\examples{
doughnut(x = c(3,5,9,12), inner.radius=0.5, col=c("red", "blue", "green", "yellow"))

}
\references{
\url{https://magesblog.com/}
}
