% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floating.floor.R
\name{floating.floor}
\alias{floating.floor}
\title{Flooring to floating values}
\usage{
floating.floor(num, digits = 1)
}
\arguments{
\item{num}{A single number or a numeric vector.}

\item{digits}{A single integer indicating the maximum number of digits required.}
}
\value{
A floored number or numeric vector.
}
\description{
Computes the floor of the given value but with any number of digits (to the closest floating number of given digits).
}
