% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.gene.name.R
\name{get.gene.name}
\alias{get.gene.name}
\title{Conversion of ENSEMBL gene IDs.}
\usage{
get.gene.name(ensembl.id, type = "gene", organism = "mmusculus")
}
\arguments{
\item{ensembl.id}{String vector of ENSEMBL genes IDs}

\item{type}{String to define the type of ENSEMBL inputs. By default \code{gene} to indicate "ensembl_gene_id". If different from "gene" it will be set to "ensembl_transcript_id_version".}

\item{organism}{String to define de organism, e.g. \code{mmusculus}, \code{hsapiens}, etc. By default \code{mmusculus}.}
}
\value{
A string vector with the corresponding gene_symbols.
}
\description{
Conversion of ENSEMBL gene IDs to gene symbols.
}
\examples{
gene_symbols =
get.gene.name(
   ensembl.id = c("ENSMUSG00000002111", "ENSMUSG00000027381"),
   type = "gene",
   organism = "mmusculus")

}
