% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.to.GOnumber.R
\name{GSEA.to.GOnumber}
\alias{GSEA.to.GOnumber}
\title{Conversion of GSEA terms into Gene Ontology numbers}
\usage{
GSEA.to.GOnumber(
  input_terms,
  input_pvalue,
  return_table = T,
  export_table = F,
  output_file_name = paste(getwd(), "GO_numbers_table.tsv", sep = "/")
)
}
\arguments{
\item{input_terms}{A character vector containing the GSEA terms to be converted.}

\item{input_pvalue}{A numeric vector containing the p-values of the GSEA terms.}

\item{return_table}{Logic value to define whether to return the resulting data.frame. By default \code{TRUE}.}

\item{export_table}{Logic value to define whether to export the resulting data.frame. By default \code{FALSE}.}

\item{output_file_name}{Path and file name of the output table if export is required. By default \code{<working.directory>/GO_numbers_table.tsv}.}
}
\value{
If required, returns a data.frame with 3 columns: GO_number, GO_annotation, p.value. This table could be directly exported.
}
\description{
Helps to convert the terms of GSEA analyses into Gene Ontology (GO) ID numbers.
}
\details{
This functions requires the package \code{GO.db}. \cr If problems are encountered during the installation see \url{https://www.biostars.org/p/50564/}.
}
