% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGVsnap.R
\name{IGVsnap}
\alias{IGVsnap}
\title{Script generator for Integrative Genomics Viewer (IGV) batch tasks.}
\usage{
IGVsnap(
  loci_vector,
  input_type,
  biomart = "ensembl",
  dataset = "mmusculus_gene_ensembl",
  reference_genome = NULL,
  fivePrime = 1000,
  threePrime = 1000,
  snap_names = NULL,
  IGV_batch_file = paste(getwd(), "/IGV_batch.txt", sep = ""),
  snap_image_format = "png",
  snap_directory = getwd(),
  maxPanelHeight = 1000,
  session = NULL,
  exit = FALSE
)
}
\arguments{
\item{loci_vector}{Either a gene name vector (e.g. \code{c("Gapdh", "Spi1", ...)}) or a regions vector (eg. \code{c('chr1:253000-256503', ...)}. All IGV formats are allowed.}

\item{input_type}{Define the input type. Allowed values are \code{genes} and \code{regions}.}

\item{biomart}{Defines the \code{biomart} parameter for \code{biomaRt} package, by default \code{ensembl}.}

\item{dataset}{Defines the \code{dataset} parameter for \code{biomaRt} package, by default \code{mmusculus_gene_ensembl}.}

\item{reference_genome}{[optional] Defines the genome to use, e.g. "mm10", "hg19", ... . By default \code{NULL}.}

\item{fivePrime}{Numeric value to define how many bases [bp] exapand from full gene position at it's 5'-end, default 1000bp.}

\item{threePrime}{Numeric value to define how many bases [bp] exapand from full gene position at it's 3'-end, default 1000bp.}

\item{snap_names}{[optional] String vector to define the names of images (without extention), by default uses \code{loci_vector}.}

\item{IGV_batch_file}{String for the batch_script_file_name/path, by default \code{<working_directory>/IGV_batch.txt}.}

\item{snap_image_format}{String to define the format of the images, e.g. "png", "jpeg", "svg", ... . By default \code{png}.}

\item{snap_directory}{String for the output directory for the snapshoots. By default <working_directory>.}

\item{maxPanelHeight}{Numeric value to define the height in pixel of the IGV pannel that will be captured on IGV.}

\item{session}{[optional] FULL path to an IGV session file (session.xml) to use for the images. By default \code{NULL}.}

\item{exit}{Logical value to indicate whether exit IGV after image capture ended. By default \code{FALSE}.}
}
\value{
Exports a .txt file ready-to-use on IGV.
}
\description{
The function builds a script file that can be run on IGV to generate multiple screenshots at specific genomic regions.
}
\details{
To run the script on IGV: Tools > Run Batch Script... > choose the .txt output file from this function. \cr For more info on how batch tasks work on IGV see: \cr \url{https://software.broadinstitute.org/software/igv/PortCommands}.
}
