% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incucyte.R
\name{incucyte}
\alias{incucyte}
\title{Incucyte analysis tool}
\usage{
incucyte(
  metadata,
  raw.data,
  comparisons = "all",
  error = "SEM",
  normalization.method = "none",
  start.hour = 0,
  plot.days = FALSE,
  show.error.ribbon = TRUE,
  show.error.bars = FALSE,
  show.points = TRUE,
  show.lines = TRUE,
  show.legend = TRUE,
  same.y.scale = TRUE,
  group.order = NULL,
  error.transparency = 0.25,
  point.size = 1,
  line.type = 1,
  line.smooth.span = 0.25,
  colors = NULL,
  skip.head.lines.in.data = 1
)
}
\arguments{
\item{metadata}{Path to a table or a data.frame with at least two columns: 'well.ID' and 'group'. The 'well.id' indicates the ID of the wells/columns in the rawData table; the 'group' indicates the name of the group at which each column belongs (wells belonging to the same column will be averaged together). A third column 'color' can be used to indicate the color; notice that each group should have the same color.}

\item{raw.data}{Path to the rawData table or a data.frame. First column must be 'Elapsed' (the time in hours) and the other columns the wells (e.g., A1, A2, B5, B8, ...).}

\item{comparisons}{List of vectors containing the groups to be used in each comparison: one vector per comparison. E.g.: list(c("group_A", "group_B"), c("group_B", "group_C", "group_E")). Default: \code{"all"}.}

\item{error}{String to indicate the error type. Possible choices: 'SEM', 'SD'. Default: \code{SEM}.}

\item{normalization.method}{String to indicate the normalization method. Possible choices: "none", "division", "subtraction". Default: \code{none}. Division: for each group the values are divided by the first timepoint.}

\item{start.hour}{Number of hours to be excluded from the calculations. Default: \code{0}.}

\item{plot.days}{Logical value to indicate whether to plotted days instead of hours. Default: \code{FALSE}.}

\item{show.error.ribbon}{Logical value to indicate whether to show the error ribbon. Default: \code{TRUE}.}

\item{show.error.bars}{Logical value to indicate whether to show the error bars. Default: \code{FALSE}.}

\item{show.points}{Logical value to indicate whether to show the individual average points at each time point. Default: \code{TRUE}.}

\item{show.lines}{Logical value to indicate whether to show the interpolation line. Default: \code{TRUE}.}

\item{show.legend}{Logical value to indicate whether to show the color legend. Default: \code{TRUE}.}

\item{same.y.scale}{Logical value to indicate whether the y-axis should show the same limit range. Default: \code{TRUE}.}

\item{group.order}{String vector with the specific order to use for the groups indicated in the metadata (factor levels). Default: \code{NULL}.}

\item{error.transparency}{Number between 0 and 1 indicating the transparency (alpha) to use for the error ribbon. Default: \code{0.25}.}

\item{point.size}{Number for the point size. Default: \code{1}.}

\item{line.type}{A vector indicating the line type to use for each group (both numeric and string values accepted). Default: \code{1} (applied to all the groups).}

\item{line.smooth.span}{Numeric value of the 'span' for the smoothing of the interpolation line. Default: \code{0.25}.}

\item{colors}{A vector indicating the color to use for each group (any R color format is accepted). Default: \code{NULL} (random colors are generated).}

\item{skip.head.lines.in.data}{Number of lines to be skipped at the beginning of the raw.data file. Default: \code{1}.}
}
\value{
The function returns a list containing:
\itemize{
  \item \code{metadata}: table used for the metadata;
  \item \code{raw.data}: tables used as raw.data;
  \item \code{analyzed.data}: data.frame with the analyzed data (means, n, SEM, SD, groups, ...);
  \item \code{normalized.data}: data.frame with the normalized data used for the plotting;
  \item \code{plot.list}: names list of plots, one plot for each group plus 'all' comparison;
  \item \code{multiplot}: a plot with all the plots in the plot.list .
 }
}
\description{
This tool generates confluency plots over time starting from incucyte data.
}
