% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.pkg.source.R
\name{install.pkg.source}
\alias{install.pkg.source}
\title{Package installer from source archive.}
\usage{
install.pkg.source(pkg.path)
}
\arguments{
\item{pkg.path}{String to define the path for the archive file to be installed.}
}
\value{
No returned value. The package required will be installed.
}
\description{
Allows the installation of R packages using the source archive file.
}
