% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect.bedtools.R
\name{intersect.bedtools}
\alias{intersect.bedtools}
\title{Intersect two or more bed files (by \code{bedtools intersect} function).}
\usage{
intersect.bedtools(
  a,
  b,
  outputFileName = paste(getwd(), "intersected.bed", sep = "/"),
  abam = FALSE,
  ubam = FALSE,
  bed = FALSE,
  wa = FALSE,
  wb = FALSE,
  loj = FALSE,
  wo = FALSE,
  wao = FALSE,
  u = FALSE,
  c = FALSE,
  C = FALSE,
  v = FALSE,
  f = NULL,
  F. = NULL,
  r = NULL,
  e = NULL,
  s = FALSE,
  S = FALSE,
  split = FALSE,
  sorted = FALSE,
  g = NULL,
  srun = FALSE,
  intersect.bedtools.command = "intersectBed",
  return.command = FALSE,
  return.bed = FALSE,
  delete.output = FALSE,
  run.command = TRUE
)
}
\arguments{
\item{a}{A single string defining the BAM/BED/GFF/VCF file “A”. Each feature in A is compared to B in search of overlaps. Use “stdin” if passing A with a UNIX pipe.}

\item{b}{A character vector with one or more BAM/BED/GFF/VCF file(s) “B”. It could be also a single string containing wildcard (*) character(s).}

\item{outputFileName}{Full path to output file name. By default \code{<working.directory>/intersected.bed}.}

\item{abam}{Logic value to define if file A is a BAM. Each BAM alignment in A is compared to B in search of overlaps. By default \code{FALSE}.}

\item{ubam}{Logic value to define if to write the output as uncompressed BAM. The default is to write compressed BAM output (\code{ubam = FALSE}).}

\item{bed}{Logic value to define whether to write output as BED when using a BAM input \code{abam = TRUE}. The default is to write output in BAM (\code{bed = FALSE}).}

\item{wa}{Logic value to define if to write the original entry in A for each overlap. By default \code{FALSE}.}

\item{wb}{Logic value to define if to write the original entry in B for each overlap. Useful for knowing what A overlaps. Restricted by -f and -r. By default \code{FALSE}.}

\item{loj}{Logic value to define if to perform a “left outer join”. That is, for each feature in A report each overlap with B. If no overlaps are found, report a NULL feature for B. By default \code{FALSE}.}

\item{wo}{Logic value to define if to write the original A and B entries plus the number of base pairs of overlap between the two features. Only A features with overlap are reported. Restricted by -f and -r. By default \code{FALSE}.}

\item{wao}{Logic value to define if to write the original A and B entries plus the number of base pairs of overlap between the two features. However, A features w/o overlap are also reported with a NULL B feature and overlap = 0. Restricted by -f and -r. By default \code{FALSE}.}

\item{u}{Logic value to define if to write original A entry once if any overlaps found in B. In other words, just report the fact at least one overlap was found in B. Restricted by -f and -r. By default \code{FALSE}.}

\item{c}{Logic value to define if to for each entry in A, report the number of hits in B while restricting to -f. Reports 0 for A entries that have no overlap with B. Restricted -f, -F, -r, and -s. By default \code{FALSE}.}

\item{C}{Logic value to define if to for each entry in A, separately report the number of overlaps with each B file on a distinct line. Reports 0 for A entries that have no overlap with B. Overlaps restricted by -f, -F, -r, and -s. By default \code{FALSE}.}

\item{v}{Logic value to define if to only report those entries in A that have no overlap in B. Restricted by -f and -r.}

\item{f}{Numeric value defining the minimum overlap required as a fraction of A. Default is 1E-9 (i.e. 1bp). By default \code{NULL}.}

\item{F.}{Numeric value defining the minimum overlap required as a fraction of B. Default is 1E-9 (i.e., 1bp). By default \code{NULL}.}

\item{r}{Numeric value defining the required reciprocal fraction of overlap for A and B. In other words, if -f is 0.90 and -r is used, this requires that B overlap at least 90\% of A and that A also overlaps at least 90\% of B. By default \code{NULL}.}

\item{e}{Numeric value defining the minimum fraction to be satisfied for A _OR_ B. In other words, if -e is used with -f 0.90 and -F 0.10 this requires that either 90\% of A is covered OR 10\% of B is covered. Without -e, both fractions would have to be satisfied. By default \code{NULL}.}

\item{s}{Logic value to define if to force “strandedness”. That is, only report hits in B that overlap A on the same strand. By default, overlaps are reported without respect to strand. By default \code{FALSE}.}

\item{S}{Logic value to define if to require different strandedness. That is, only report hits in B that overlap A on the _opposite_ strand. By default, overlaps are reported without respect to strand. By default \code{FALSE}.}

\item{split}{Logic value to define if to treat “split” BAM (i.e., having an “N” CIGAR operation) or BED12 entries as distinct BED intervals. By default \code{FALSE}.}

\item{sorted}{Logic value to define, for very large B files, if to invoke a “sweeping” algorithm that requires position-sorted input. When using -sorted, memory usage remains low even for very large files. By default \code{FALSE}. It is possible to sort a bed file on terminal by \code{(sort -k1,1 -k2,2n unsorted.bed > sorted.bed)} or by the function \link{sort.bed}.}

\item{g}{Specify a genome file the defines the expected chromosome order in the input files for use with the -sorted option. By default \code{NULL}.}

\item{srun}{Logic value to define whether the command should be run in \code{srun} mode. By default \code{FALSE}.}

\item{intersect.bedtools.command}{String to define the command to use to recall the \code{bedtools intersect} function. An example: "/home/user/anaconda3/bin/intersectBed". By default \code{"intersectBed"}.}

\item{return.command}{Logic value to define whether to return the string corresponding to the command for bedtools. By default \code{FALSE}.}

\item{return.bed}{Logic value to define whether to return the resulting bed as data.frame. By default \code{FALSE}. Parameter not active when inputs are bam files.}

\item{delete.output}{Logic value to define whether to delete the exported intersected bed file. By default \code{FALSE}. Parameter active only when \code{return.bed = TRUE}. Useful when is sufficient to get the result as a data.frame without saving it.}

\item{run.command}{Logic value to define whether to run the the command line on system terminal and generate the bed resulting from the intersection. By default \code{TRUE}.}
}
\value{
The function generates the files indicated by the output parameters. If required the command line used and/or the resulting intersected bed file. If both outputs are required, the output will be a named list with two values: "command" and "intersected.bed".
}
\description{
This function runs a command line that uses \code{bedtools intersect} to intersect one or more .bed files.
}
\details{
To know more about the \code{bedtools intersect} function see the package manual at the following link: \cr \url{https://bedtools.readthedocs.io/en/latest/content/tools/intersect.html}.
}
\examples{
intersect.bedtools(a = bed_file1.bed,
                   b = c("bed_file2.bed", "bed_file3.bed"),
                   wb = TRUE,
                   intersect.bedtools.command = "/home/user/anaconda3/bin/intersectBed")

intersect.bedtools(a = bed_file1.bed,
                   b = c("bed_file2.bed", "bed_file3.bed"),
                   wa = TRUE,
                   return.bed = TRUE,
                   delete.output = T,
                   intersect.bedtools.command = "/home/user/anaconda3/bin/intersectBed")

}
