% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.color.R
\name{is.color}
\alias{is.color}
\title{is.color}
\usage{
is.color(x)
}
\arguments{
\item{x}{A string vector.}
}
\value{
A logical vector of the same length of \code{x}.
}
\description{
Function to define if each element of a string vector is an R-supported color string.
}
