% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molarity.to.mass.R
\name{molarity.to.mass}
\alias{molarity.to.mass}
\title{Get solvent volume to make a solution with a given amount of a compound.}
\usage{
molarity.to.mass(
  final_concentration,
  final_concentration_unit = "M",
  final_volume,
  final_volume_unit = "mL",
  MW
)
}
\arguments{
\item{final_concentration}{Numeric value for the final concentration wanted.}

\item{final_concentration_unit}{String to define the unit of the final concentration wanted. Available units are: "M", "mM", "uM", "nM", "pM", "fM". By default "M".}

\item{final_volume}{Numeric value for the final volume wanted.}

\item{final_volume_unit}{String to define the unit of the volume. Available units are: "L", "mL", "uL". By default "mL".}

\item{MW}{Numeric value for the Molecular Weigth (MW) of the compound expressed in g/mol.}
}
\value{
It returns a string with the mass of compound to use.
}
\description{
Given a specific volume of solution wanted calculates the mass of solute necessary to obtain a certain final molarity concentration.
}
\examples{
molarity.to.mass(final_concentration = 5, final_volume = 10, MW = 215)

}
