% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.df.col.R
\name{move.df.col}
\alias{move.df.col}
\title{Function to change easily the order of specific columns in a data.frame.}
\usage{
move.df.col(data.frame, move.command)
}
\arguments{
\item{data.frame}{An input data.frame.}

\item{move.command}{A string containing the moving command. The command is formed as follows: "columnA movingCommand columnB". The basic options are: "first", "last", "before", "after". Compounded moves must be separated by a semicolon. Example: \code{"g first; a last; e before c"}.}
}
\value{
It returns the original data.frame but with the columns moved as demanded.
}
\description{
Allows to change the position of a column in a data.frame using other columns as reference.
}
\examples{
new.mtcars = move.df.col(mtcars, "mpg last")

new.mtcars = move.df.col(mtcars, "wt before carb")

new.mtcars = move.df.col(mtcars, "am before carb; cyl first")

}
\references{
\url{https://stackoverflow.com/questions/3369959/moving-columns-within-a-data-frame-without-retyping}
}
