% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.bigwig.mean.R
\name{multi.bigwig.mean}
\alias{multi.bigwig.mean}
\title{Multi-bigWig average tool}
\usage{
multi.bigwig.mean(
  sample.config,
  output.dir = file.path(getwd(), "merged_bigWigs"),
  bin.size = 50,
  merged.suffix = paste0("_merged_bs", format(ceiling(bin.size), scientific = FALSE)),
  out.file.format = "bigwig",
  number.of.processors = 4,
  bigwigCompare = "bigwigCompare",
  return.log = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{sample.config}{A data.frame or a string indicating the path to a table. The table/data.frame must contain in the first column the file list, while the second column the group to which belong each file. Files belonging to the same group will be merged together. The group name corresponds to the name of the corresponding output file.}

\item{output.dir}{String of the path to the output directory. Default: \code{file.path(getwd(), "merged_bigWigs")}.}

\item{bin.size}{Size of the bins in base-pairs (bp). Default: \code{50}.}

\item{merged.suffix}{Suffix to use for the resulting output files: <group.id><suffix><extension>. Default: \code{paste0("_merged_bs", format(ceiling(bin.size), scientific = FALSE))}.}

\item{out.file.format}{Output file format. Possible choices: "bigWig", "bw", "bedGraph", "bdg". Default: \code{"bigwig"}.}

\item{number.of.processors}{Number of CPUs to use. Default: \code{4}.}

\item{bigwigCompare}{String of the path to the \code{bigwigCompare} tool. Default: \code{"~/.conda/envs/snakepipes/b9364eb954bd13fe4e5f104fa8a286e6/bin/bigwigCompare"}.}

\item{return.log}{Logic value to define whether the log list of the \code{bigwigCompare} steps performed by 'group.id'. Default: \code{TRUE}.}

\item{verbose}{Logic value to define whether the function should print messages. Default: \code{TRUE}.}
}
\value{
A list of vectors with the commands used to run \code{bigwigCompare} at each step. Each element is one group ID.
}
\description{
This tools uses \code{bigwigCompare} (see details) to perform the average between multiple bigWig files in a multi-step process.
}
\details{
For details on \code{bigwigCompare} visit: \url{https://deeptools.readthedocs.io/en/develop/content/tools/bigwigCompare.html}
}
