% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.check.R
\name{pkg.check}
\alias{pkg.check}
\title{Check package installation.}
\usage{
pkg.check(package, archive)
}
\arguments{
\item{package}{A single string indicating the name of the package to check.}

\item{archive}{A single string indicating the type of archive. Possible values "CRAN" and "bioconductor" (not case sensitive). Parameter without default..}
}
\value{
If the pkg is not already installed it will be installed.
}
\description{
Function to check if a package is installed. It works with bioconductor or CRAN packages.
}
\examples{
pkg.check("ggplot2", "cran")

pkg.check("biomaRt", "bioconductor")

}
