% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.density.summary.R
\name{plot.density.summary}
\alias{plot.density.summary}
\title{Plot the distribution of overall NGS density at specific regions from deepTools matrices.}
\usage{
\method{plot}{density.summary}(
  matrix.file,
  plot.by.group = T,
  missing.data.as.zero = NULL,
  sample.names = NULL,
  region.names = NULL,
  signal.type = "mean",
  linear = F,
  error.type = "sem",
  show.mean = T,
  mean.error.type = "se",
  mean.color = "blue",
  mean.symbol.shape = 20,
  mean.symbol.size = 1,
  show.stat.multiplot = T,
  stat.method = "wilcox.test",
  stat.paired = F,
  stat.labels.format = "p.signif",
  stat.hide.ns = T,
  stat.p.levels = list(cutpoints = c(0, 1e-04, 0.001, 0.01, 0.05, 1), symbols =
    c("****", "***", "**", "*", "ns")),
  title = NULL,
  x.lab = NULL,
  y.lab = NULL,
  x.labs.angle = 0,
  dodge.width = 1,
  border.width = 0.5,
  transparency = 0.5,
  subset.range = NULL,
  y.lim = NULL,
  y.identical.auto = T,
  y.ticks.interval = NULL,
  y.digits = 1,
  text.size = 12,
  legend.position = c(0.2, 0.85),
  colors = c("#00A5CF", "#F8766D", "#AC88FF", "#E08B00", "#00BA38", "#BB9D00",
    "#FF61C9", "gray30"),
  n.row.multiplot = 1,
  multiplot.export.file = NULL,
  real.width.single.violinplot = 1,
  real.height.single.violinplot = 3.5,
  by.row = TRUE,
  print.multiplot = F
)
}
\arguments{
\item{matrix.file}{A single string indicating a full path to a matrix.gz file generated by \code{deepTools/computeMatrix} or by \link{computeMatrix.deeptools}, or a list generated by the function \link{read.computeMatrix.file}.}

\item{plot.by.group}{Logical value to define whether plot by group of regions or by sample. By default \code{TRUE}.}

\item{missing.data.as.zero}{Logical value to define whether treat missing data as 0. If set as \code{FALSE} missing data will be converted to \code{NA} and will be excluded from the computations of the signal. By default \code{TRUE}.}

\item{sample.names}{Samples names could be defined by a string vector. If set as \code{NULL} sample names will be get automatically by the matrix file. By default \code{NULL}. \cr Example: \code{c("sample1", "sample2", "sample3")}}

\item{region.names}{Region names could be defined by a string vector. If set as \code{NULL} sample names will be get automatically by the matrix file. By default \code{NULL}. \cr Example: \code{c("regionA", "regionB")}}

\item{signal.type}{String indicating the signal to be computed and plotted. Available parameters are "mean", "median" and "sum". By default "mean".}

\item{linear}{Logical value to define whether the plots should show the score in linear scale. By default \code{FALSE}.}

\item{error.type}{String indicating the type of error to be computed and that will be available in the output data.table. Available parameters are "sem" and "sd", standard error mean and standard deviation respectively. By default "sem". Parameter considered only when \code{show.mean = TRUE)}.}

\item{show.mean}{Logical value to define whether the mean value should be shown as a symbol on the plots. By default \code{TRUE}.}

\item{mean.error.type}{String indicating the type of error for the mean to be computed. Available parameters are "se", "sd" and, "none". Respectively standard error, standard deviation, and no error plotted. By default "se". Parameter considered only when \code{show.mean = TRUE)}.}

\item{mean.color}{A single string expressing an R-supported color for the mean symbol. By default \code{"blue"}.}

\item{mean.symbol.shape}{A numeric value or string defining the shape for the mean symbol. By default \code{20}.}

\item{mean.symbol.size}{A numeric value defining the size of the mean symbol. By default \code{1}.}

\item{show.stat.multiplot}{Logical value to define if to add to the plot the statistical comparisons of the means for the groups present in the multiplot. By default \code{TRUE}. All possibile comparisons will be performed.}

\item{stat.method}{A single string defining the method to use for the statistical comparisons. By default \code{"wilcox.test"}. Available options: "t.test" "wilcox.test" "anova" "kruskal.test".}

\item{stat.paired}{Logical value to define if the statistical comparisons should be performed paired. By default \code{"FALSE"}. Notice that to allow a paired comparison the number of data should be the same in the two groups compared, so in the most of the cases non applicable to the comparisons between two regions. Used only in \code{"t.test"} and \code{"wilcox.test"} methods.}

\item{stat.labels.format}{A single string indicating the format of the p-value to show for the statistical comparisons. By default \code{"p.signif"}. Available options: "p.format" (normal p-value), "p.signif" (significance stars), "p.adj" (p-value adjusted).}

\item{stat.hide.ns}{Logical value indicating if the NS ("Not Significant") comparisons should be shown or not. By default \code{TRUE}.}

\item{stat.p.levels}{A list containing the p-values levels/thresholds in the following format (default): \code{list(cutpoints = c(0, 0.0001, 0.001, 0.01, 0.05, 1), symbols = c("****", "***", "**", "*", "ns"))}. In other words, we use the following convention for symbols indicating statistical significance:
\itemize{
  \item \code{ns}: p > 0.05
  \item \code{*} p <= 0.05
  \item \code{**} p <= 0.01
  \item \code{***} p <= 0.001
  \item \code{****} p <= 0.0001
 }}

\item{title}{Title of each plot could be defined by a string vector. If set as \code{NULL} titles will be generated automatically. By default \code{NULL}. \cr Example: \code{c("Title1", "Title2")}}

\item{x.lab}{Single string or string vector to define the X-axis label for all the plots. By default \code{NULL}, the label will be defined automatically.}

\item{y.lab}{Single string or string vector to define the Y-axis label for all the plots. By default \code{NULL}, the label will be defined automatically.}

\item{x.labs.angle}{A single numeric value indicating the degrees of rotation of the category labels in the X-axis. By default \code{0}, horizontal without rotation.}

\item{dodge.width}{Numeric value defining the width of each single violin plot. By default \code{1}.}

\item{border.width}{Numeric value to define the border width for all the plots. By default \code{0.5}.}

\item{transparency}{A numeric value to define the fraction of transparency of the plots fill (0 = transparent, 1 = full). By default \code{0.5}.}

\item{subset.range}{A numeric vector indicating the range to which restrict the analyses (eg. c(-150, 250)). In the case of "scale-region" mode, the range is represented by (-upstream | 0 | body_length | body_length+downstream).By default \code{NULL}: the whole region is considered.}

\item{y.lim}{List of numeric vectors with two elements each to define the range of the Y-axis. To set only one side use NA for the side to leave automatic. If only one range is given this one will be applied to all the plots. By default \code{NULL}, the range will be defined automatically. \cr Example \code{list(c(0, 20), c(NA, 30), c(0, NA), c(NA, NA))}.,}

\item{y.identical.auto}{Logical value to define whether use the same Y-axis range for all the plots automatically depending on the values. Not used when \code{y.lim} is not \code{NULL}. By default \code{TRUE}.}

\item{y.ticks.interval}{A number indicating the interval/bin spacing two ticks on the Y-axis. By default \code{NULL}: ticks are assigned automatically. Active only when \code{y.identical.auto = TRUE} and \code{y.lim != NULL}.}

\item{y.digits}{A numeric value to define the number of digits to use for the y.axis values. By default \code{1} (eg. 1.5).}

\item{text.size}{Numeric value to define the size of the text for the labels of all the plots. By default 12.}

\item{legend.position}{Any ggplot supported value for the legend position (eg. "none, "top", "bottom", "left", "right", c(fraction.x, fraction.y)). By default \code{c(0.2, 0.85)}.}

\item{colors}{Vector to define the line and error area colors. If only one value is provided it will applied to all the samples/groups. If the number of values is lower than the the required one, a random set of colors will be generated. All standard R.colors values are accepted. By default \code{c("#00A5CF", "#F8766D", "#AC88FF", "#E08B00", "#00BA38", "#BB9D00", "#FF61C9", "gray30")}.}

\item{n.row.multiplot}{Numeric value to define the number of rows in the final multiplot.}

\item{multiplot.export.file}{If a string with the name of a PDF file is provided the multiplot will be exported. By default \code{NULL}.}

\item{real.width.single.violinplot}{Numeric value, in inches, to define the real width (not precise) of each single violin plot in the multiplot exported, if required. By default 1 inch.}

\item{real.height.single.violinplot}{Numeric value, in inches, to define the real height (not precise) of each single violin plot in the multiplot exported, if required. By default 3.5 inches.}

\item{by.row}{Logical value to define whether the plots should be arranged by row. By default \code{TRUE}.}

\item{print.multiplot}{Logical value to define whether to print the multiplot once generated. By default \code{FALSE}.}
}
\value{
The functions returns a list containing:
\itemize{
  \item \code{data.table} with the computed values used for the plot;
  \item \code{metadata} table with the information obtained from the matrix_file.gz;
  \item \code{plot.list} with a plot for each list element;
  \item \code{multiplot} with the image of all the plots together;
  \item \code{summary.plot.samples} with a plot showing the scores of all regions per each sample;
  \item \code{summary.plot.regions} with a plot showing the scores of all samples per each region;
  \item \code{means.comparisons} table with the statistical means comparisons (when \code{show.stat.multiplot = TRUE}, otherwise a string is returned).
 }
}
\description{
Computes the score of each element in a list of regions and generates violins plots with percentiles and the mean (optional) for each sample/region. It uses as input a score matrix computed by deeptools's computeMatrix function or by \link{computeMatrix.deeptools} from this package.
}
\details{
To know more about the deepTools's function \code{computeMatrix} see the package manual at the following link: \cr \url{https://deeptools.readthedocs.io/en/develop/content/tools/computeMatrix.html}.
}
