% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pStars.R
\name{pStars}
\alias{pStars}
\title{P-value significance stars definer.}
\usage{
pStars(p.value, one = 0.05, two = 0.01, three = 0.001, four = 1e-04)
}
\arguments{
\item{p.value}{A single numeric value indicating the p-value to evaluate.}

\item{one}{A numeric value to define the p-value threshold for the first level of significance (*). By default 0.05.}

\item{two}{A numeric value to define the p-value threshold for the second level of significance (**). By default 0.01.}

\item{three}{A numeric value to define the p-value threshold for the third level of significance (***). By default 0.001.}

\item{four}{A numeric value to define the p-value threshold for the fourth level of significance (****). By default 0.0001.}
}
\value{
It returns a string with the corresponding level of significance: \code{NS}, \code{*}, \code{**}, \code{***}, \code{****}.
}
\description{
Converts a p-value score in equivalent stars of significance.
}
\examples{
significance = pStars(0.002)

require(dplyr)
data.frame =
   data.frame \%>\%
   mutate(p.stars = sapply(data.frame$p.value.column, pStars))

}
