% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qPCR.results}
\alias{qPCR.results}
\title{qPCR RNA expression results example}
\format{
A data frame with 117 rows and 35 variables. Three of these columns are required to run \link{\code{qPCR.rna.exp}}:
\describe{
  \item{\code{Sample Name}}{Name of the samples/conditions}
  \item{\code{Target Name}}{The target genes to quantify}
  \item{\code{CT}}{Values of the cycle detected at a given threshold}
}
}
\source{
Simulated data
}
\usage{
qPCR.results
}
\description{
Simulation of appliedBiosystem qPCR results
}
\keyword{datasets}
