% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCR.rna.exp.R
\name{qPCR.rna.exp}
\alias{qPCR.rna.exp}
\title{qPCR RNA expression analyses tool.}
\usage{
qPCR.rna.exp(
  results.file,
  housekeeping.genes = NULL,
  max.delta.reps = 0.5,
  reference.sample = NULL,
  exlude.houskeeping.FC = TRUE,
  exclude.samples = NULL,
  fix.y.axis = FALSE,
  x.labels.rotation = 45,
  text.size = 3,
  results.sheet.position = 3,
  rows.to.skip = 44,
  file.header = TRUE,
  file.tail = TRUE,
  samples.order = NULL,
  ignore.reps.errors = FALSE
)
}
\arguments{
\item{results.file}{String indicating the full path to the results excel file or a data.frame containing at least the following columns: 'Sample Name', 'Target Name', 'CT'.}

\item{housekeeping.genes}{String vector with the list of genes that have to be used as target genes. By default \code{NULL}: an error message is printed.}

\item{max.delta.reps}{Numeric value indicating the maximum difference among replicate Ct. Default value: \code{0.5}.}

\item{reference.sample}{Single string indicating the name of the sample to use as reference for the computation of the FoldChanges. By default \code{NULL}: the first sample in the order is used as reference.}

\item{exlude.houskeeping.FC}{Logic value to indicate whether the housekeeping genes should be excluded in the FoldChanges plots. By default \code{TRUE}.}

\item{exclude.samples}{String vector indicating the samples that should be exuded in the expression and FoldChange plots. By default \code{NULL}.}

\item{fix.y.axis}{Logic value indicating whether the y-axis of the plots should be kept fixed among all the genes. By default \code{FALSE}.}

\item{x.labels.rotation}{Numeric value indicating the degrees of x-axis's labels rotation. By default \code{45}.}

\item{text.size}{Numeric value to indicate the size of the text for the number above the bars. Default \code{3}.}

\item{results.sheet.position}{Numeric value indicating the position of the results sheet in the excel file. by default \code{3}.}

\item{rows.to.skip}{How many rows must be skipped before to read the excel file. By default \code{44}.}

\item{file.header}{Logic value to indicate whether the results excel file contains an header. By default \code{TRUE}.}

\item{file.tail}{Logic value to indicate whether the results excel file contains extra rows at the end of the results. By default \code{TRUE}.}

\item{samples.order}{A string vector indicating all the samples in order. This order will be used to order the samples in the plots. By default \code{NULL}: the reference sample will be the first, the other will be kept in the order available in the results table.}

\item{ignore.reps.errors}{Logic value to define whether the difference between the Ct in replicates should be ignored: all the values are kept.}
}
\value{
The function returns a list containing:
\itemize{
  \item \code{original.table}: a data.frame containing the original results table;
  \item \code{reshaped.table}: a data.frame with the original results reorganized for the analyses;
  \item \code{reshaped.table.cleaned}: the reshaped data.frame upon filtering of the CT values (if required);
  \item \code{reps.validation.plot}: a plot representing a table with the differences two-by-two of the technical replicates (facet_wrapped by gene) where the cells have a red background if the difference is greater than the `max.delta.reps` value;
  \item \code{analyzed.data}: a named list of data.frames, one for each housekeeping gene and one for the foldChange mean of all housekeeping normalization, containing the normalized expression scores and the FoldChanges over the reference sample;
  \item \code{expression.plots}: a named list of plots, one for each housekeeping gene, showing the gene expression histograms (facet_wrapped by gene);
  \item \code{foldChange.plots}: a named list of plots, one for each housekeeping gene and one for the foldChange mean of all housekeeping normalization, showing the FoldChange expression over the reference Sample (facet_wrapped by gene).
 }
}
\description{
Allows to easily analyse qPCR RNA expression data, including: technical replicates verification, gene expression normalization to housekeeping genes and FoldChanges over reference sample computation.
}
