% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.computeMatrix.file.R
\name{read.computeMatrix.file}
\alias{read.computeMatrix.file}
\title{\code{computeMatrix} *.gz file reader}
\usage{
read.computeMatrix.file(matrix.file)
}
\arguments{
\item{matrix.file}{A string indicating indicating the full path to the matrix.file.gz generated by deepTools/computeMatrix function or by \link{computeMatrix.deeptools}.}
}
\value{
The functions returns a named list containing:
\itemize{
  \item \code{metadata} data.frame with the information gotten from the matrix_file.gz;
  \item \code{matrix.data} data.frame with the scores gotten from;
  \item \code{original.file.path} with full path to the original matrix_file.gz.
 }
 This list can be passed as it is to the function \link{plot.density.profile}.
}
\description{
The function reads a matrix.file.gz generated by deepTools/computeMatrix function or by \link{computeMatrix.deeptools}. The value can be passed to \link{plot.density.profile} function.
}
