% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.samples.computeMatrix.R
\name{reorder.samples.computeMatrix}
\alias{reorder.samples.computeMatrix}
\title{Sample reorderer computeMatrix file}
\usage{
\method{reorder}{samples.computeMatrix}(
  matrix.file,
  new.sample.order = NULL,
  reordered.matrix.path = gsub(".gz", "_reordered.gz", matrix.file),
  ignore.header.error = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{matrix.file}{String with the full path to a deeptools computeMatrix \code{.gz} file.}

\item{new.sample.order}{String vector with the sample labels in the order in which they should appear in the matrix. By default {NULL}, which returns a message with the sample labels in the original order. Further, it returns the vector with the original sample order.}

\item{reordered.matrix.path}{String with full path to for the file of the reorder matrix (.gz). By default the output name will be \code{<original.matrix.name>_reordered.gz}.}

\item{ignore.header.error}{Logical value to indicate whether the error of sample_label reassignment in the header should be ignored. The plotted labels can be changed during the plotting.}

\item{verbose}{Logical value to indicate whether the final message should be printed. By default \code{TRUE}.}
}
\value{
The output is a computeMatrix file (.gz format) with the samples chucks re-shuffled to be in the order provided by the user.
}
\description{
A tool to reorder the samples in a computeMatrix file avoiding the re-computation of the latter.
}
