% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restriction.sites.to.bed.R
\name{restriction.sites.to.bed}
\alias{restriction.sites.to.bed}
\title{Generator of a bed file for enzymatic restriction sites.}
\usage{
restriction.sites.to.bed(
  cut_positions,
  chromosome,
  genome_start,
  return_bed = TRUE,
  export_bed_file = FALSE,
  output_file_name = paste(getwd(), "restriction_positions.bed", sep = "/"),
  enzyme_cut_length = 4,
  include_region_description = TRUE,
  region_name = "site",
  append = FALSE
)
}
\arguments{
\item{cut_positions}{A numeric vector with the list of the restriction/cut positions.}

\item{chromosome}{Chromosome number of the region analyzed.}

\item{genome_start}{Start position on the genome of the region analyzed.}

\item{return_bed}{Logic value to define if to return the bed as data.frame. By default \code{TRUE}.}

\item{export_bed_file}{Logic value to define if to export the resulting .bed file. By default \code{FALSE}.}

\item{output_file_name}{String corresponding to the path to the exported .bed file. By default \code{"<working.directory>/restriction_positions.bed"}.}

\item{enzyme_cut_length}{Numeric value to define the length of cut of the restriction enzyme. By default 4.}

\item{include_region_description}{Logic value to define whether to include a fourth column containing the region name define by the parameter \code{region_description}. By default \code{TRUE}.}

\item{region_name}{Regions base name. Automatically it will be added a number to the base name. By default \code{"site"}, the resulting regions will be: site_1, site_2, ... .}

\item{append}{Logic value to define if to append the result to the file. By default \code{FALSE}, the file will be overwritten.}
}
\value{
If required, it will be returned a classic bed file (chr start end [name]) with the regions centered on the cut position in the genome.
}
\description{
The function allows to create a bed file that can be added on IGV both as regions and track. It will show the restriction sites of a sequences if starting from the cut positions depending on sequence lenght. Chromosome, start and end of the input sequence are required.
}
\details{
To map the positions of restriction enzymes it is possible to use \url{http://restrictionmapper.org/} with the option \code{Map (version 3)}.
}
\examples{
restriction.sites.to.bed(cut_positions = c(230, 235, 1250, 36),
                         chromosome = 10,
                         genome_start = 1205126,
                         region_name = "EcoRI_cut_site")

}
