% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNAseq}
\alias{RNAseq}
\title{RNA-seq example}
\format{
A data frame with 300 rows and 7 variables:
\describe{
  \item{\code{geneName}}{genes symbols}
  \item{\code{baseMean}}{The average of the normalized count values, dividing by size factors, taken over all samples}
  \item{\code{log2FC}}{the log2 value of the Fold Change expression between two conditions}
  \item{\code{lfcSE}}{log2 Fold Change standard error (SE)}
  \item{\code{stat}}{Wald statistic}
  \item{\code{pvalue}}{Wald test p-value}
  \item{\code{padj}}{BH adjusted p-values}
}
}
\source{
Simulated data
}
\usage{
RNAseq
}
\description{
Dummy example of a DESeq2 result for differential expression analysis on RNA-seq data
}
\keyword{datasets}
