% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{Rseb-options}
\alias{Rseb-options}
\title{Package options}
\description{
The \code{Rseb} package uses the following global options
(set via \code{\link[base:options]{options()}}):

\describe{
  \item{\code{Rseb.update_check}}{
    Logical. Whether Rseb checks for package updates.
    Default is \code{TRUE}.
  }
}
}
\keyword{internal}
