% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.bed.R
\name{sort.bed}
\alias{sort.bed}
\title{Sorter function for .bed files.}
\usage{
\method{sort}{bed}(
  bed,
  bed.header = FALSE,
  sep = "\\t",
  return.bed = TRUE,
  export.file.name = NULL,
  export.header = FALSE,
  unique.regions = TRUE
)
}
\arguments{
\item{bed}{Two options are possible: \cr - String with the path to a .bed file; \cr - Data.frame corresponding to a bed file format (all the columns and their names will be kept).}

\item{bed.header}{Logic value to define whether the .bed file contains an header or not. By default \code{FALSE}.}

\item{sep}{String containing the separator character for a .bed file. By default \code{"\t"}.}

\item{return.bed}{Logic value to define if to return the bed as a data.frame. By default \code{TRUE}. Only unique rows are kept.}

\item{export.file.name}{Optional: string to define the path to the file to be exported, if required. By default \code{NULL}, not exported.}

\item{export.header}{Logic value to define whether the header should be exported in the sorted bed file. By default \code{FALSE}.}

\item{unique.regions}{Logic value to indicate whether the output bed must contain unique regions. By default \code{TRUE}.}
}
\value{
If required, returns a data.frame corresponding to the sorted .bed file.
}
\description{
Sorts .bed files by chromosome and position.
}
\details{
The function keeps only unique rows. \cr To get more information about the bed file format see the following page: \cr \url{https://genome.ucsc.edu/FAQ/FAQformat.html#format1}.
}
