% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_packages.R
\name{store_packages}
\alias{store_packages}
\title{Stores the information of installed packages in a .rda file.}
\usage{
store_packages(output_directory = getwd())
}
\arguments{
\item{output_directory}{Path to the directory in which export the .rda file. By default \code{<working.directory>}.}
}
\value{
Nothing is returned. An .rda file will be exported at the \code{output_directory} indicated.
}
\description{
Saves the list of all the installed packages in a .rda file. This file can be used to restore the packages from a computer to another or after installation of a new R version by the function \code{\link{restore_packages}} of this package.
}
