% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform.y.axis.R
\name{uniform.y.axis}
\alias{uniform.y.axis}
\title{Plot Y-axis uniforming}
\usage{
uniform.y.axis(plot.list, y.min = TRUE, y.max = TRUE, ticks.each = NULL)
}
\arguments{
\item{plot.list}{A single plot or a list of plots.}

\item{y.min}{Either a logical value to define whether uniform the lower limit or a numeric value defining the lower limit. By default \code{TRUE}.}

\item{y.max}{Either a logical value to define whether uniform the upper limit or a numeric value defining the upper limit. By default \code{TRUE}.}

\item{ticks.each}{Numeric value to define every how much should be placed a tick. By default \code{NULL}, ticks will be placed automatically.}
}
\value{
Returns a plot list (or a single plot when only one input plot is provided) equivalent to the input list provided by the user in which the Y-axis of all the plots will be uniformed.
}
\description{
Takes a list of ggplot2 plots, compares their Y-axis ranges and applies the highest/lowest limits to each plot in order to uniform all the plots. It can be used also to set the ticks step (to just change the breaks set all parameters as \code{FALSE}).
}
