% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pkgs.R
\name{update_pkgs}
\alias{update_pkgs}
\title{function to automatically update the R packages.}
\usage{
update_pkgs(ask = FALSE)
}
\arguments{
\item{ask}{Logical indicating whether to ask the user to select packages before they are downloaded and installed, or the character string \code{"graphics"}, which brings up a widget to allow the user to (de-)select from the list of packages which could be updated. (The latter value only works on systems with a GUI version of \code{select.list}, and is otherwise equivalent to \code{ask = TRUE}). By default \code{FALSE}.}
}
\value{
Nothing. The packages will be updated.
}
\description{
Automatically updates the R packages from CRAN and BioConductor repositories.
}
\examples{
update_pkgs()

}
