% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcano}
\alias{volcano}
\title{Volcano plot generator for RNA-seq data.}
\usage{
volcano(
  log2FC_data,
  padj_data,
  FC_t = 1.5,
  p_t = 0.05,
  FC_unresponsive_rigth = 1.1,
  FC_unresponsive_left = 1/FC_unresponsive_rigth,
  x_ends = NULL,
  y_min = 0,
  y_max = NULL,
  left_label = "UP",
  right_label = "DOWN",
  unresponsive_label = "NoResp",
  null_label = "NULL",
  names = as.character(c(1:length(log2FC_data))),
  left_names = FALSE,
  right_names = FALSE,
  padding = FALSE,
  names_size = 10,
  print_plot = F,
  left_color = "#00BA38",
  right_color = "#F8766D",
  unresponsive_color = "#00A5CF",
  null_color = "gray30",
  point_size = 0.5,
  legend = TRUE,
  legend_title = "Expression status",
  x_label = "log2(fold change expression)",
  y_label = "-log10(p-value adjusted)",
  title = "Volcano plot",
  sub_title = NULL,
  add_threshold_lines = T,
  threshold_line_color = "gray70",
  threshold_line_type = "dotted",
  font_family = "Helvetica",
  font_size = 12
)
}
\arguments{
\item{log2FC_data}{Numeric vector containing the log2(FoldChange) values of each gene.}

\item{padj_data}{Numeric vector of p-values. Use of adjusted p-values is recommended.}

\item{FC_t}{Value of the threshold to use for the fold change expression to define differentially expressed genes, expressed as linear value. By default 1.5 and by consequence 1/1.5.}

\item{p_t}{Value of the threshold to use for the p-values to define differentially expressed genes, expressed as linear value. By default 0.05.}

\item{FC_unresponsive_rigth}{Value of the threshold to use for the fold change expression to define unresponsive genes when \code{FC > 1}, expressed as linear value. By default 1.1. If \code{NULL} it will be calculated symmetrically from \code{FC_NoResp_left} as 1/\code{FC_NoResp_left}.}

\item{FC_unresponsive_left}{Value of the threshold to use for the fold change expression to define unresponsive genes when \code{FC < 1}, expressed as linear value. By default \code{1/FC_unresponsive_rigth}. If \code{NULL} it will be calculated symmetrically from \code{FC_NoResp_rigth} as 1/\code{FC_NoResp_rigth}.}

\item{x_ends}{Numeric positive value to define manually the range of the X-axis: it will be calculated as \code{c(-x_ends, x_ends)}, for this reason the plot will be symmetrical. By default \code{NULL}, the range is assigned automatically and the plot can be asymmetrical.}

\item{y_min}{Numeric value for the minimum value of the Y-axis. By default 0. Set it to \code{NULL} for automatic computation.}

\item{y_max}{Numeric value for the maximum value of the Y-axis. By default \code{NULL}.}

\item{left_label}{String to indicate the label to use for the set of genes in the left side of the graph (those with \code{FoldChange < 1/FC_t} and \code{p.value < p_t}. By default \code{"UP"}.}

\item{right_label}{String to indicate the label to use for the set of genes in the right side of the graph (those with \code{FoldChange > FC_t} and \code{p.value < p_t}. By default \code{"DOWN"}.}

\item{unresponsive_label}{String to indicate the label to use for the set of unresponsive genes (those with \code{FC_unresponsive_left < FoldChange < FC_unresponsive_rigth} and \code{p.value > p_t}. By default \code{"NoResp"}.}

\item{null_label}{String to indicate the label to use for the set of null genes (those with \code{1/FC_t < FoldChange < FC_t} and \code{p.value < p_t}. By default \code{"NULL"}.}

\item{names}{String vector with the names to be plotted if required, eg. gene names. By default \code{as.character(c(1:length(log2FC_data)))}.}

\item{left_names}{Logic value to indicate if to print the set of differentially expressed genes in the left side of the graph (those with \code{FoldChange < 1/FC_t} and \code{p.value < p_t}. By default \code{FALSE}.}

\item{right_names}{Logic value to indicate if to print the set of differentially expressed genes in the right side of the graph (those with \code{FoldChange > FC_t} and \code{p.value < p_t}. By default \code{FALSE}.}

\item{padding}{Logic value to indicate if to plot the padding around the names of genes. By default \code{FALSE}.}

\item{names_size}{Numeric value to define de size of the point names size. By default 10.}

\item{print_plot}{Logic value to define whether to print the volcano plot once created. By default \code{FALSE}.}

\item{left_color}{String to indicate the color to use for the set of genes in the left side of the graph (those with \code{FoldChange < 1/FC_t} and \code{p.value < p_t}. By default \code{"#00BA38"}, a green.}

\item{right_color}{String to indicate the color to use for the set of genes in the right side of the graph (those with \code{FoldChange > FC_t} and \code{p.value < p_t}. By default \code{"#F8766D"}, a pink/red.}

\item{unresponsive_color}{String to indicate the color to use for the set of unresponsive genes (those with \code{FC_unresponsive_left < FoldChange < FC_unresponsive_rigth} and \code{p.value > p_t}. By default \code{"#00A5CF"}, a light blue.}

\item{null_color}{String to indicate the color to use for the set of null genes (those with \code{1/FC_t < FoldChange < FC_t} and \code{p.value < p_t}. By default \code{"gray30"}, a dark gray.}

\item{point_size}{Numeric value to define de size of the points. By default 0.5.}

\item{legend}{Logic value to define if to print the legend. By default \code{TRUE}.}

\item{legend_title}{A string to indicate the label of the legend title. By default \code{"Expression status"}.}

\item{x_label}{A string to indicate the X-axis label. By default \code{"log2(fold change expression)"}.}

\item{y_label}{A string to indicate the Y-axis label. By default \code{"-log10(p-value adjusted)"}.}

\item{title}{A string to indicate the title of the plot. By default \code{"Volcano plot"}.}

\item{sub_title}{A string to indicate the subtitle of the plot. By default \code{NULL}, no subtitle is written.}

\item{add_threshold_lines}{Logic value to define if lines for the thresholds, both FC and p.value, should be plotted. By default \code{TRUE}.}

\item{threshold_line_color}{String to define the color of the threshold lines. By default \code{"gray70"}}

\item{threshold_line_type}{String or numeric value to define the threshold lines type. Both numeric and string standard R codes are accepted. By default \code{"dotted"}, equivalent to \code{2}.}

\item{font_family}{String to define the font family to use in the plot writings. By default \code{"Helvetica"}.}

\item{font_size}{Numeric value to define the font size. By default 12.}
}
\value{
A plot in ggplot2 format.
}
\description{
Generates a volcano plot in order to visualize the differentially expressed genes. The plot is highly customizable.
}
