% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actualize.R
\name{actualize}
\alias{actualize}
\title{\code{Rseb} updates verification}
\usage{
actualize(update = TRUE, verbose = TRUE, force = FALSE)
}
\arguments{
\item{update}{Logical value to define whether update the \code{Rseb} package. By default \code{TRUE}.}

\item{verbose}{Logical value to define whether print messages. By default \code{TRUE}.}

\item{force}{Logical value to define whether to force the installation of \code{Rseb} even though already up-to-date. Parameter passed to \code{devtools::install_github()}. By default \code{FALSE}.}
}
\value{
Warnings and/or messages. Installation of the latest version of \code{Rseb} if required.
}
\description{
It verifies if \code{Rseb} is up-to-date and installs it when required.
}
\details{
This function will check for internet availability.
}
