% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.density.profile.R
\name{plot.density.profile}
\alias{plot.density.profile}
\title{Plot of NGS density signal at specific regions from deepTools matrices.}
\usage{
\method{plot}{density.profile}(
  matrix.file,
  plot.by.group = T,
  missing.data.as.zero = NULL,
  sample.names = NULL,
  region.names = NULL,
  signal.type = "mean",
  error.type = "sem",
  plot.error = T,
  error.transparency = 0.125,
  title = NULL,
  x.lab = NULL,
  y.lab = NULL,
  line.type = "solid",
  line.width = 0.5,
  x.lim = NULL,
  y.lim = NULL,
  y.identical.auto = T,
  y.ticks.number = 5,
  text.size = 12,
  plot.vertical.lines = T,
  write.reference.points = T,
  colors = c("#00A5CF", "#F8766D", "#AC88FF", "#E08B00", "#00BA38", "#BB9D00",
    "#FF61C9", "gray30"),
  n.row.multiplot = 1,
  export.multiplot = F,
  multiplot.export.file = paste(getwd(), "/multiplot.", Sys.Date(), "_", gsub(pattern =
    ":", replacement = ".", x = format(Sys.time(), "\%X")), ".pdf", sep = ""),
  real.width.single.plot = 2.5,
  real.height.single.plot = 3,
  print.multiplot = F
)
}
\arguments{
\item{matrix.file}{A single string indicating a full path to a matrix.gz file generated by \code{deepTools/computeMatrix} or by \link{computeMatrix.deeptools}, or a list generated by the function \link{read.computeMatrix.file}.}

\item{plot.by.group}{Logic value to define whether plot by group of regions or by sample. By default \code{TRUE}.}

\item{missing.data.as.zero}{Logic value to define whether treat missing data as 0. If set as \code{FALSE} missing data will be converted to \code{NA} and will be excluded from the computations of the signal. By default \code{TRUE}.}

\item{sample.names}{Samples names could be defined by a string vector. If set as \code{NULL} sample names will be get automatically by the matrix file. By default \code{NULL}. \cr Example: \code{c("sample1", "sample2", "sample3")}}

\item{region.names}{Region names could be defined by a string vector. If set as \code{NULL} sample names will be get automatically by the matrix file. By default \code{NULL}. \cr Example: \code{c("regionA", "regionB")}}

\item{signal.type}{String indicating the signal to be computed and plotted. Available parameters are "mean", "median" and "sum". By default "mean".}

\item{error.type}{String indicating the type of error to be computed and plotted. Available parameters are "sem" and "sd", standard error mean and standard deviation respectively. By default "sem". Parameter considered only when \code{plot.error = TRUE)}.}

\item{plot.error}{Logic value to define whether to plot the error around the signal. By default \code{TRUE}.}

\item{error.transparency}{Numeric value to define the alpha/transparency of the error. By default 0.125. Parameter considered only when \code{plot.error = TRUE)}.}

\item{title}{Title of each plot could be defined by a string vector. If set as \code{NULL} titles will be generated automatically. By default \code{NULL}. \cr Example: \code{c("Title1", "Title2")}}

\item{x.lab}{Single string to define the X-axis label for all the plots. By default \code{NULL}, the label will be defined automatically.}

\item{y.lab}{Single string to define the Y-axis label for all the plots. By default \code{NULL}, the label will be defined automatically.}

\item{line.type}{Vector to define each line type. Both numeric and string codes are accepted. If only one element is given this will be applied to all the lines. By default "solid". \cr Example 1: \code{c("solid", "dashed")}. \cr Example 2: \code{c(1, 2)}}

\item{line.width}{Numeric value to define the line width for all the plots. By default 0.5.}

\item{x.lim}{List of numeric vectors with two elements each to define the range of the X-axis. To set only one side use NA for the side to leave automatic. If only one range is given this one will be applied to all the plots. By default \code{NULL}, the range will be defined automatically. \cr Example \code{list(c(0, 20), c(NA, 30), c(0, NA), c(NA, NA))}.,}

\item{y.lim}{List of numeric vectors with two elements each to define the range of the Y-axis. To set only one side use NA for the side to leave automatic. If only one range is given this one will be applied to all the plots. By default \code{NULL}, the range will be defined automatically. \cr Example \code{list(c(0, 20), c(NA, 30), c(0, NA), c(NA, NA))}.,}

\item{y.identical.auto}{Logical value to define whether use the same Y-axis range for all the plots automatically depending on the values. Not used when \code{y.lim} is not \code{NULL}. By default \code{TRUE}.}

\item{y.ticks.number}{Define the number of ticks to display in the Y-axis. By default 5. Active only when \code{y.identical.auto = TRUE}.}

\item{text.size}{Numeric value to define the size of the text for the labels of all the plots. By default 12.}

\item{plot.vertical.lines}{Logic value to define whether to plot a dashed gray vertical line in correspondence of the reference points of each plot. By default \code{TRUE}.}

\item{write.reference.points}{Logic value to define whether to indicate the reference points on each plot. Applied only when \code{x.lim} is \code{NULL}. By default \code{TRUE}.}

\item{colors}{Vector to define the line and error area colors. If only one value is provided it will applied to all the samples/groups. If the number of values is lower than the the required one, a random set of colors will be generated. All standard R.colors values are accepted. By default \code{c("#00A5CF", "#F8766D", "#AC88FF", "#E08B00", "#00BA38", "#BB9D00", "#FF61C9", "gray30")}.}

\item{n.row.multiplot}{Numeric value to define the number of rows in the final multiplot.}

\item{export.multiplot}{Logic value to define whether to export the multiplot generated by the function. By default \code{FALSE}.}

\item{multiplot.export.file}{Name of the PDF file of the multiplot to be exported when \code{export.multiplot = T}. \cr By default "/working_directory/multiplot_current.date_current.time.pdf".}

\item{real.width.single.plot}{Numeric value, in inches, to define the real width of each plot in the multiplot exported, if required. By default 2.5 inches.}

\item{real.height.single.plot}{Numeric value, in inches, to define the real height of each plot in the multiplot exported, if required. By default 3 inches.}

\item{print.multiplot}{Logic value to define whether to print the multiplot once created. By default \code{FALSE}.}
}
\value{
The functions returns a list containing:
\itemize{
  \item \code{data.table} with the computed values used for the plot;
  \item \code{metadata} table with the information gotten from the matrix_file.gz;
  \item \code{plot.list} with a plot for each list element;
  \item \code{multiplot} with the image of all the plots together.
 }
}
\description{
Plots the density profile of NGS data signals, using as input a score matrix computed by deeptools's computeMatrix function or by \link{computeMatrix.deeptools} from this package.
}
\details{
To know more about the deepTools's function \code{computeMatrix} see the package manual at the following link: \cr \url{https://deeptools.readthedocs.io/en/develop/content/tools/computeMatrix.html}.
}
\examples{
plot.density.profile(
  matrix.file = "/path.to/matrix.file.gz", plot.by.group = TRUE,
  missing.data.as.zero = NULL, sample.names = NULL, region.names = NULL,
  signal.type = "mean", error.type = "sem", plot.error = TRUE,
  error.transparency = 0.125, title = NULL, x.lab = NULL, y.lab = NULL,
  line.type = "solid", line.width = 0.5, x.lim = NULL, y.lim = NULL,
  y.identical.auto = TRUE, y.ticks.number = 5, text.size = 12,
  plot.vertical.lines = TRUE, colors = c("red", "blue", "#00BA38"),
  n.row.multiplot = 1, export.multiplot = FALSE,
  multiplot.export.file = "/path.to/multiplot.pdf",
  real.width.single.plot = 2.5, real.height.single.plot = 3,
  print.multiplot = FALSE)

}
